Moving storage provider tests into a common test fixture
