+++ .gitignore
@@ -32,3 +32,6 @@ com_crashlytics_export_strings.xml
 # generated by npm
 node_modules/
 npm-debug.log
+
+# Application-Generated Files
+test-sqlite3.db
