+++ package.json
@@ -20,6 +20,7 @@
     "jit-grunt": "^0.9.0",
     "mocha": "^2.1.0",
     "mysql": "^2.5.4",
+    "q": "^1.1.2",
     "sqlite3": "^3.0.4",
     "time-grunt": "^1.0.0"
   }
